+++ index.html
@@ -22,7 +22,7 @@
   <body>
     <div id="flat-chat"></div>
     <script type="module">
-      import { FormExtension } from "./src/extensions/MyExtension.tsx";
+      import { FormExtension } from "./src/index.tsx";
       const PROJECT_ID = import.meta.env.VITE_VOICEFLOW_PROJECT_ID;
       const ENV = import.meta.env.NODE_ENV;
